Refactor: move `Recruitment` to `users` namespace.
