+++ carp.studies.core/src/commonMain/kotlin/dk/cachet/carp/studies/application/ParticipantService.kt
@@ -6,9 +6,9 @@ import dk.cachet.carp.common.data.Data
 import dk.cachet.carp.common.data.input.InputDataType
 import dk.cachet.carp.common.ddd.ApplicationService
 import dk.cachet.carp.common.ddd.IntegrationEvent
-import dk.cachet.carp.studies.domain.ParticipantGroupStatus
 import dk.cachet.carp.studies.domain.users.AssignParticipantDevices
 import dk.cachet.carp.studies.domain.users.Participant
+import dk.cachet.carp.studies.domain.users.ParticipantGroupStatus
 
 
 /**
