+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/domain/users/ParticipantRepositoryTest.kt
@@ -1,7 +1,6 @@
 package dk.cachet.carp.studies.domain.users
 
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.studies.domain.Recruitment
 import dk.cachet.carp.test.runSuspendTest
 import kotlin.test.*
