+++ typescript-declarations/tests/carp.studies.core.ts
@@ -29,7 +29,7 @@ import getAssignedDeviceRoles = dk.cachet.carp.studies.domain.users.deviceRoles_
 import DeanonymizedParticipant = dk.cachet.carp.studies.domain.users.DeanonymizedParticipation
 import Participant = dk.cachet.carp.studies.domain.users.Participant
 import StudyOwner = dk.cachet.carp.studies.domain.users.StudyOwner
-import ParticipantGroupStatus = dk.cachet.carp.studies.domain.ParticipantGroupStatus
+import ParticipantGroupStatus = dk.cachet.carp.studies.domain.users.ParticipantGroupStatus
 import StudyDetails = dk.cachet.carp.studies.domain.StudyDetails
 import StudyStatus = dk.cachet.carp.studies.domain.StudyStatus
 import StudyServiceRequest = dk.cachet.carp.studies.infrastructure.StudyServiceRequest
